# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Gatekeeper Firewall API
  version: "1.0.0"
  description: |
    REST API for the Gatekeeper network firewall control plane.
    All mutation endpoints require API key authentication via `X-API-Key` header.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - apiKey: []
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Zone:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        interface:
          type: string
        network_cidr:
          type: string
        trust_level:
          type: string
          enum: [none, low, medium, full]
        description:
          type: string
      required: [name]
    Alias:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [host, network, port, mac, nested, external_url]
        description:
          type: string
        members:
          type: array
          items:
            type: string
      required: [name, type]
    Policy:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        default_action:
          type: string
          enum: [allow, deny, reject, log]
        rules:
          type: array
          items:
            $ref: "#/components/schemas/Rule"
      required: [name, default_action]
    Rule:
      type: object
      properties:
        id:
          type: integer
        policy_id:
          type: integer
        order:
          type: integer
        src_alias:
          type: string
        dst_alias:
          type: string
        protocol:
          type: string
        ports:
          type: string
        action:
          type: string
          enum: [allow, deny, reject, log]
        log:
          type: boolean
        description:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
paths:
...
  /api/v1/zones:
    get:
      summary: List all zones
      responses:
        "200":
          description: Zone list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Zone"
    post:
      summary: Create a zone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Zone"
      responses:
        "201":
          description: Zone created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/zones/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get a zone by name
      responses:
        "200":
          description: Zone details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
        "404":
          description: Not found
    put:
      summary: Update a zone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Zone"
      responses:
        "200":
          description: Zone updated
    delete:
      summary: Delete a zone
      responses:
        "200":
          description: Zone deleted
  /api/v1/aliases:
    get:
      summary: List all aliases
      responses:
        "200":
          description: Alias list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alias"
    post:
      summary: Create an alias
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Alias"
      responses:
        "201":
          description: Alias created
...
  /api/v1/policies:
    get:
      summary: List policies
      responses:
        "200":
          description: Policy list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Policy"
    post:
      summary: Create policy with rules
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Policy"
      responses:
        "201":
          description: Policy created
...
  /api/v1/assign:
    post:
      summary: Assign device to profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ip:
                  type: string
                mac:
                  type: string
                hostname:
                  type: string
                profile:
                  type: string
                profile_id:
                  type: integer
              required: [ip]
      responses:
        "201":
          description: Device assigned
...